+++ blog/changelog.md
@@ -1,3 +1,8 @@
 ---
 title: Changelog
 ---
+
+## 2022-03-22
+
+- Add "Compare Projects" tools
+- Add "Try It Yourself" blog